Added interface between internal and external bus
The interface between internal and external busses is implemented with
this commit. I implemented a DRAM similarly to how we implemented memory
in assignment 1. The DRAM access protocol has been shown to work with a
single read and single write. I still have to add functionality for
burst operations. This will be done when adding cache access.

Danilo